Update test dependencies to Flask==1.0.2 and Werkzeug==0.14.1
During #3741, we updated Flask and Werkzeug in the securedrop-
app-code package dependencies. However, these same dependencies
were not updated in the test dependencies file, which include
Flask and Werkzeug (due to the pip-compile logic we have). This
meant that in dev, the older versions of these deps were used.
Even in staging, the older versions of these deps was used due
to the fact that we install the test dependencies directly
in the staging VMs _after_ the securedrop-app-code package is
installed.